requirements.txt
pillow>=9.0.0
pytesseract>=0.3.8
mss>=6.1.0
pywin32>=303
interception-python>=0.1.0
numpy==1.26.4
rapidfuzz==3.6.1
opencv-python==4.9.0.80
pynput==1.7.6 